Changed message when user_finder returns nil
When the user_finder returns nil it should not use a custom error
message. The user should be able to define this error message. By
letting the message return nil the user can set a default message in the
warden options and have that returned if the user is not found.